Extend cigar count to uint32_t

+++ src/main/native/smithwaterman/smithwaterman_common.h
@@ -41,7 +41,7 @@ typedef struct dnaSeqPair
         int16_t max_j;
         int16_t *btrack;
         char *cigar;
-        int16_t cigarCount;
+        uint32_t cigarCount;
         int16_t alignmentOffset;
 }SeqPair;
